LPS-55505 move to module
