+++ modules/portal/portal-scripting-ruby/src/com/liferay/portal/scripting/ruby/internal/RubyExecutor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.scripting.ruby;
+package com.liferay.portal.scripting.ruby.internal;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
